胡志成提交

+++ src/main/java/cn/gson/his/model/pojos/hzc/BedEntity.java
@@ -11,7 +11,7 @@ public class BedEntity {
     private int bedId; //床位编号
     private Integer depaId; //所属科室编号
     private Integer bedIs; //床位状态
-    private Long bedCost; //床位费用
+    private Double bedCost; //床位费用
 
     //科室对象
     private Department department;
